+++ web_manage_schedule_exam/admin/manage_student.php
@@ -86,7 +86,7 @@
                                 echo '<td>' . $khoa . '</td>';
                                 echo '<td>' . $gioi_tinh . '</td>';
                                 echo '<td id="update-icon"><a href="update_student.php?msv=' . urlencode($msv) . '"><i class="fa-solid fa-pen-to-square"></i></a></td>';
-                                echo '<td id="delete-icon"><a href="delete.php?msv=' . urlencode($msv) . '"><i class="fa-solid fa-trash-can"></i></a></td>';   
+                                echo '<td id="delete-icon"><a href="delete.php?msv=' . urlencode($msv) . '" onclick="return confirmDelete(\'' . $msv . '\')"><i class="fa-solid fa-trash-can"></i></a></td>';
                                 echo '</tr>';  
                             }
                         } else {
@@ -109,7 +109,7 @@
                 if ($result_sinh_vien->num_rows > 0) {
                     echo '<div class="selected-box">
                             <span> Chọn tất cả <input type="checkbox" onClick="toggle(this)" /></span>
-                                <button type="submit" name="delete_multiple_student"><i class="fa-solid fa-trash-can"></i></button>
+                                <button type="submit" name="delete_multiple_student" onclick="return confirmDeleteAll()"><i class="fa-solid fa-trash-can"></i></button>
                             </form>
                         </div>';
                 }
